aegnt/requirements.txt
google-adk
google-generativeai
httpx
python-dotenv
fastapi
uvicorn
